move registry into own files
