+++ inlang/packages/paraglide-js/src/compiler/output-structure/message-modules.ts
@@ -1,6 +1,6 @@
 import type { ProjectSettings } from "@inlang/sdk";
 import type { CompiledBundleWithMessages } from "../compileBundle.js";
-import { createRuntime } from "../runtime.js";
+import { createRuntime } from "../runtime/create-runtime.js";
 import { createRegistry } from "../registry.js";
 import { jsIdentifier } from "../../services/codegen/identifier.js";
 import { escapeForSingleQuoteString } from "../../services/codegen/escape.js";
